lib.yang: factor out lib.yang.ctype
